Image module API reorganization (issue aspectron/thorium#16)
Moved functions to generate PNG, JPEg and BMP from bitmap images into free functions in encoder.hpp/cpp to hide dependencies from zlib, libpng and libjpeg
Extracted rescaler class into a separate translation unit

+++ src/image.cpp
@@ -29,7 +29,7 @@ void image_uninstall(Handle<Value> library)
 {
 }
 
-uint64_t bitmap::total_memory_ = 0;
+size_t bitmap::total_memory_ = 0;
 
 enum grid_type { GRID_LINE, GRID_1, GRID_2 };
